requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
entsoe-py>=0.5.10
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
ruff>=0.1.0

# Optional: for enhanced data handling
python-dateutil>=2.8.0
